TVB-2687: Move operation related methods to operation_service and fix review so it works for simulations also
