+++ framework_tvb/tvb/core/adapters/abcadapter.py
@@ -53,7 +53,6 @@ from tvb.core.adapters import constants
 from tvb.core.entities.generic_attributes import GenericAttributes
 from tvb.core.entities.load import load_entity_by_gid
 from tvb.core.neocom import h5
-from tvb.core.neotraits._h5core import ViewModelH5
 from tvb.core.neotraits.h5 import H5File
 from tvb.core.utils import date2string, LESS_COMPLEX_TIME_FORMAT
 from tvb.core.entities.storage import dao
@@ -523,41 +522,6 @@ class ABCAdapter(object):
             msg = "Could not load Adapter Instance for Stored row %s" % stored_adapter
             LOGGER.exception(msg)
             raise IntrospectionException(msg)
-
-    def review_operation_inputs(self, operation):
-        """
-        :returns: a list with the inputs from the parameters list that are instances of DataType,\
-            and a dictionary with all parameters which are different than the declared defauts
-        """
-        changed_attr = {}
-        inputs_datatypes = []
-        view_model = self.load_view_model(self, operation)
-        form_model = self.get_view_model_class()()
-        form_fields = self.get_form_class()().fields
-
-        for field in form_fields:
-            if not isinstance(field, TraitDataTypeSelectField):
-                attr_form = getattr(form_model, field.name)
-                attr_vm = getattr(view_model, field.name)
-                if attr_vm != attr_form:
-                    changed_attr[field.label] = attr_vm
-            else:
-                attr_vm = getattr(view_model, field.name)
-                data_type = ABCAdapter.load_entity_by_gid(attr_vm)
-                inputs_datatypes.append(data_type)
-
-        return inputs_datatypes, changed_attr
-
-    def load_view_model(self, adapter_instance, operation):
-        storage_path = FilesHelper().get_project_folder(operation.project, str(operation.id))
-        input_gid = json.loads(operation.parameters)['gid']
-        view_model_class = adapter_instance.get_view_model_class()
-        view_model = view_model_class()
-        h5_path = h5.path_for(storage_path, ViewModelH5, input_gid)
-        h5_file = ViewModelH5(h5_path, view_model)
-        h5_file.load_into(view_model)
-        return view_model
-
 
 @add_metaclass(ABCMeta)
 class ABCAsynchronous(ABCAdapter):
